% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidsHistogram.R
\name{AminoAcidsHistogram}
\alias{AminoAcidsHistogram}
\title{Make a histogram plot out of an amino acid string}
\usage{
AminoAcidsHistogram(AminoAcids = "")
}
\arguments{
\item{AminoAcids}{String of amino acids}
}
\value{
ggplot2::ggplot
}
\description{
Make a histogram plot out of an amino acid string
}
\examples{
AminoAcidsHistogram('QHHPKHVAFKLGFAVGSTKSYHKFLSGHNV')
}
