% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_into_aa.R
\name{codons_into_aa}
\alias{codons_into_aa}
\title{Conversion from codons to amino acids}
\usage{
codons_into_aa(codons)
}
\arguments{
\item{codons}{table that helps translating codons into amino-acids}
}
\value{
the amino-acid sequence
}
\description{
Conversion from codons to amino acids
}
