% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonsToAA.R
\name{codonsToAA}
\alias{codonsToAA}
\title{Convert input codons via codon table to Amino acids}
\usage{
codonsToAA(codons)
}
\arguments{
\item{codons}{Input condons generated}
}
\value{
Return Amino acids from codon
}
\description{
Convert input codons via codon table to Amino acids
}
