% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_T_U.R
\name{convert_T_U}
\alias{convert_T_U}
\title{convert T into U}
\usage{
convert_T_U(DNA_sequence)
}
\arguments{
\item{DNA_sequence}{Original DNA sequence that we wish to convert}
}
\value{
It returns the associated RNA sequence
}
\description{
convert T into U
}
