% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaToRna.R
\name{dnaToRna}
\alias{dnaToRna}
\title{DNA to RNA Transcription}
\usage{
dnaToRna(dnaSeq)
}
\arguments{
\item{dnaSeq}{The DNA sequence to be transcribed}
}
\value{
The RNA transcript of the DNA sequence
}
\description{
DNA to RNA Transcription
}
