% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Extracts codons from a RNA sequence}
\usage{
extract_codons(rna_seq, start = 1)
}
\arguments{
\item{rna_seq}{RNA sequence}

\item{start}{Start position of the first codon}
}
\value{
A vector of codons extracted from the RNA sequence starting from the start position provided
}
\description{
Extracts codons from a RNA sequence
}
\examples{
rna_seq <- "AUGGUGAUG"
start <- 1
extract_codons(rna_seq, start) # [1] "AUG" "GUG" "AUG"
}
