% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stop_start.R
\name{find_stop_start}
\alias{find_stop_start}
\title{Find start and stop codons}
\usage{
find_stop_start(dna_seq)
}
\arguments{
\item{dna_seq}{A string DNA sequence}
}
\value{
Index positions for start and stop codons in the sequence
}
\description{
Find start and stop codons
}
\examples{
dna_sequence <- "ATGCGTAAATGCGTAGATGCCTGA"
codon_positions <- find_stop_start(dna_sequence)
print(codon_positions)
}
