% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_5.R
\name{function_5}
\alias{function_5}
\title{A function which takes a data frame of codons and counts them before plotting in a column plot (idk?)}
\usage{
function_5(codon_table)
}
\arguments{
\item{codon_table}{a dataframe of codons}
}
\value{
returns a column plot of codon counts
}
\description{
A function which takes a data frame of codons and counts them before plotting in a column plot (idk?)
}
