% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_5.R
\name{plot_aas}
\alias{plot_aas}
\title{A function which takes a data frame of codons and counts them before plotting in a column plot (idk?)}
\usage{
plot_aas(amino_acids)
}
\arguments{
\item{amino_acids}{a string of amino acids in single letter format}
}
\value{
returns a column plot of codon counts
}
\description{
A function which takes a data frame of codons and counts them before plotting in a column plot (idk?)
}
