% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreqAA.R
\name{plotFreqAA}
\alias{plotFreqAA}
\title{Frequency plot of amino acids in sequence}
\usage{
plotFreqAA(AASeq)
}
\arguments{
\item{AASeq}{Protein sequence}
}
\value{
Plot of the frequency for amino acids
}
\description{
Frequency plot of amino acids in sequence
}
