% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna1}
\alias{random_dna1}
\title{Creating random DNA}
\usage{
random_dna1(random_dna2)
}
\arguments{
\item{random_dna2}{desired lenght of DNA to create}
}
\value{
Returns a random DNA-like sequence with the input lenght
}
\description{
Creating random DNA
}
\examples{
10 is the input by the user, the function will create a one continue
string of random DNA-like with a lenght of 10 nucleotide, example of an output: ACCTGAAGCT

}
