% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDNASeq.R
\name{randomDNAseq}
\alias{randomDNAseq}
\title{randomDNAseq}
\usage{
randomDNAseq(seqLength)
}
\arguments{
\item{seqLength}{a number}
}
\value{
Returns a random dna sequence as a string of same length as the input number
}
\description{
randomDNAseq
}
