% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_codons.R
\name{read_codons}
\alias{read_codons}
\title{Codon reader}
\usage{
read_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{String of DNA sequence.}

\item{start}{Intereger. Starting position in sequence for reading codons (default=1).}
}
\value{
Character vector containing the codons.
}
\description{
Codon reader
}
\examples{
my_sequence <- "ATGCTAGCGTACTAG"
read_codons(my_sequence)
read_codons(my_sequence, start = 2)

}
