% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_nuc.R
\name{sample_nuc}
\alias{sample_nuc}
\title{Sample nucleotides}
\usage{
sample_nuc(nr_nuc)
}
\arguments{
\item{nr_nuc}{Integer - the number of nucleotides to sample}
}
\value{
String - A sequence of nucleotides the length of the input parameter
}
\description{
This function generates a random sequence of nucleotides.
}
\examples{
sample_nuc(1)
sample_nuc(10)
}
