% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqToCodon.R
\name{seqToCodon}
\alias{seqToCodon}
\title{Convert input sequence to codons}
\usage{
seqToCodon(inputSeq, start = 1)
}
\arguments{
\item{inputSeq}{This is the input sequence}

\item{start}{Position to start in the sequence}
}
\value{
Function returns codons from input sequence
}
\description{
Convert input sequence to codons
}
