% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_DNA.R
\name{translate_DNA}
\alias{translate_DNA}
\title{Translate DNA to RNA}
\usage{
translate_DNA(DNA_string)
}
\arguments{
\item{DNA_string}{DNA string}
}
\value{
A RNA string - the DNA string but with U's in stead of T's
}
\description{
Translate DNA to RNA
}
\examples{
translate_DNA('AG')
translate_DNA('AGTCTT')
}
