% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_codon.R
\name{transscript_aa_sequence}
\alias{transscript_aa_sequence}
\title{Transscript Amino acid Sequence (Function_4)}
\usage{
transscript_aa_sequence(codons)
}
\arguments{
\item{codons}{<- A string of RNA codons}
}
\value{
A vector of transscripted amino acids in pairs of 3 in their
        single letter form
}
\description{
Transscript Amino acid Sequence (Function_4)
}
\examples{
aa_sequence("AUGUUGCCGCAACGG")
aa_sequence("AUGUAA")
}
